tech_comp_final/backend/requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
pydantic==2.12.3
httpx==0.28.1
aiohttp==3.13.1
beautifulsoup4==4.14.2
lxml==5.4.0
python-docx==1.2.0
trafilatura==2.0.0
python-dateutil==2.9.0
pandas==2.2.2
redis==5.0.1
celery==5.3.4
requests
pdfplumber
python-docx
matplotlib
pandas
rapidfuzz
sentence-transformers
torch
transformers
python-multipart
python-dotenv
